[2019 day 18] Speed up solution
The previous solution walked the grid on each search, wasting a bunch of
time. Instead a three-phase approach is used to parse the grid, generate
the graph of reachable keys, and finally walk the graph.

This dropped the debug runtime down from ~200 seconds to ~3 seconds and
release from ~10 seconds to ~300 milliseconds.